requirements.txt
# Flask Web Application Dependencies

Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0

# Google Cloud
google-cloud-firestore==2.14.0

# Spotify & Genius
lyricsgenius==3.0.1
spotipy==2.23.0

# Utilities
requests==2.31.0
